requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.3.0
annotated-types==0.7.0
anthropic==0.34.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autopep8==2.3.1
azure-ai-documentintelligence==1.0.0b3
azure-core==1.30.2
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
black==24.4.2
blinker==1.8.2
bs4==0.0.2
build==1.2.1
cachetools==5.4.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
docx2txt==0.8
faiss-cpu==1.8.0.post1
fastapi==0.112.1
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-auth==2.34.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
googletrans==4.0.0rc1
grpcio==1.65.5
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.0.0
importlib_resources==6.4.3
isodate==0.6.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==30.1.0
langchain==0.2.15
langchain-anthropic==0.1.23
langchain-chroma==0.1.2
langchain-cli==0.0.29
langchain-community==0.2.15
langchain-core==0.2.37
langchain-ollama==0.1.0
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langgraph==0.2.16
langgraph-checkpoint==1.0.3
langgraph-checkpoint-sqlite==1.0.1
langserve==0.2.2
langsmith==0.1.86
libcst==1.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.0
onnxruntime==1.19.0
openai==1.41.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
posthog==3.5.0
proto-plus==1.24.0
protobuf==4.21.12
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pydantic==1.10.13
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.8.0
PyPika==0.48.9
pyproject-toml==0.0.10
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
setuptools==72.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.31
sse-starlette==1.8.2
starlette==0.38.2
streamlit==1.36.0
sympy==1.13.2
tavily-python==0.4.0
telegram==0.0.1
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.23.2
uvloop==0.20.0
watchfiles==0.23.0
websocket-client==1.8.0
websockets==12.0
wheel==0.44.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.0
